requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-genai
easyocr
pillow
numpy
opencv-python
streamlit
requests